Separate your classes into different files
